api/requirements.txt
fastapi
pydantic
sentiment-analysis-spanish
scikit-learn
httpx
python-dotenv
pytest
